modalとcardをコンポーネント化
