+++ menu-list-frontend/pages/_app.tsx
@@ -1,8 +1,11 @@
 import '../styles/globals.css'
 import type { AppProps } from 'next/app'
+import Card  from "./components/card";
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
+  return (
+    <Card />
+  )
 }
 
 export default MyApp
